Restructure server code.
