+++ tic_tac_toe_server/utils/middleware.js
@@ -1,0 +1,8 @@
+
+const unknownEndpoint = (request, response) => {
+  response.status(404).send({error: 'unkown endpoint'});
+}
+
+module.exports = {
+  unknownEndpoint
+}
